// src/tcgdb/modules/tcgdb/pokemon/price/change/weekly/dto/tcgdb.pokemon.price.change.weekly.dto.ts
export class TCGdbPokemonPricesChangeWeeklyDTO {
    tcgdbPokemonPricesChangeWeekly: TCGdbPokemonPriceChangeWeeklyDTO[];
}

export class TCGdbPokemonPriceChangeWeeklyDTO {
    tcgdbPokemonPriceChangeWeeklyId: string;
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeWeeklyTCGPlayerId: number;
    tcgdbPokemonPriceChangeWeeklySetCode: string;
    tcgdbPokemonPriceChangeWeeklyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeWeeklyLowPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyLowPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeWeeklyMidPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyMidPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeWeeklyHighPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyHighPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeWeeklyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklySubTypeName: string;
}

export class CreateTCGdbPokemonPriceChangeWeeklyDTO {
    tcgdbPokemonCardId: string;
    tcgdbPokemonPriceChangeWeeklyTCGPlayerId: number;
    tcgdbPokemonPriceChangeWeeklySetCode: string;
    tcgdbPokemonPriceChangeWeeklyCurrentLowPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousLowPrice: number;
    tcgdbPokemonPriceChangeWeeklyLowPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyLowPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentMidPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousMidPrice: number;
    tcgdbPokemonPriceChangeWeeklyMidPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyMidPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentHighPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousHighPrice: number;
    tcgdbPokemonPriceChangeWeeklyHighPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyHighPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklyCurrentMarketPrice: number;
    tcgdbPokemonPriceChangeWeeklyPreviousMarketPrice: number;
    tcgdbPokemonPriceChangeWeeklyMarketPriceDifference: number;
    tcgdbPokemonPriceChangeWeeklyMarketPricePercentage: number;
    tcgdbPokemonPriceChangeWeeklySubTypeName: string;
}